mono_camera_control example: update to use `with`
